moved code to the SearchBarController
