+++ src/main/java/com/teama/controllers/MainScreenController.java
@@ -107,31 +107,30 @@ public class MainScreenController implements Initializable {
 
     private MapSubsystem mapSubsystem;
 
-    String[] words = {"Fenwood Road",
-            "Schlagler Stairs",
-            "Information Desk",
-            "Information Desk",
-            "Elevator S G",
-            "BTM Security Desk",
-            "Neuro Testing Waiting Area",
-            "Infusion Waiting Area",
-            "Schlagler Innovation Lobby",
-            "Test Lobby",
-            "Test Area1",
-            "Test Area2",
-            "Test Area3",
-            "Test Desk",
-            "Test Front Desk",
-            "Test Info Desk",
-            "Test Desk2",
-            "Test Desk3",
-            "Test Desk4",
-            "Test Desk5",
-            "Test Stairs2",
-
-    };
-    Set<String> possibleWordSet = new HashSet<>();
-    private AutoCompletionBinding<String> autoCompletionBinding;
+//    String[] words = {"Fenwood Road",
+//            "Schlagler Stairs",
+//            "Information Desk",
+//            "Information Desk",
+//            "Elevator S G",
+//            "BTM Security Desk",
+//            "Neuro Testing Waiting Area",
+//            "Infusion Waiting Area",
+//            "Schlagler Innovation Lobby",
+//            "Test Lobby",
+//            "Test Area1",
+//            "Test Area2",
+//            "Test Area3",
+//            "Test Desk",
+//            "Test Front Desk",
+//            "Test Info Desk",
+//            "Test Desk2",
+//            "Test Desk3",
+//            "Test Desk4",
+//            "Test Desk5",
+//            "Test Stairs2",
+//    };
+//    Set<String> possibleWordSet = new HashSet<>();
+//    private AutoCompletionBinding<String> autoCompletionBinding;
 
 
     @FXML
@@ -160,7 +159,7 @@ public class MainScreenController implements Initializable {
         drawer.setSidePane();
         hamOpnsTran = new HamburgerBackArrowBasicTransition(hamburgerButton);
 
-        Collections.addAll(possibleWordSet, words);
+//        Collections.addAll(possibleWordSet, words);
 
 
         //String[] possibleWords = {"test", "test2"};
@@ -180,19 +179,19 @@ public class MainScreenController implements Initializable {
         PathfindingController pathfinding = new PathfindingController(MapSubsystem.getInstance(), map, areaPane, floorButtonBox);
         SearchBarController searchBarController = new SearchBarController(searchBar, searchButton, mapSubsystem);
 
-        searchBar.setEditable(true);
-        autoCompletionBinding = TextFields.bindAutoCompletion(searchBar.getEditor(), possibleWordSet);
-        //TextFields.bindAutoCompletion(searchBar.getEditor(), possibleWordSet);
-
-        searchBar.getEditor().setOnKeyPressed((KeyEvent e)->{
-            switch(e.getCode()){
-                case ENTER:
-                    learnWord(searchBar.getEditor().getText());
-                    break;
-                default:
-                    break;
-            }
-        });
+//        searchBar.setEditable(true);
+//        //autoCompletionBinding = TextFields.bindAutoCompletion(searchBar.getEditor(), possibleWordSet);
+//        //TextFields.bindAutoCompletion(searchBar.getEditor(), possibleWordSet);
+//
+//        searchBar.getEditor().setOnKeyPressed((KeyEvent e)->{
+//            switch(e.getCode()){
+//                case ENTER:
+//                    learnWord(searchBar.getEditor().getText());
+//                    break;
+//                default:
+//                    break;
+//            }
+//        });
 
         // Zoom in and out using plus and minus keys
         mapScroll.onKeyTypedProperty().set((KeyEvent event) -> {
@@ -287,14 +286,14 @@ public class MainScreenController implements Initializable {
         });
     }
 
-    private void learnWord(String text) {
-        possibleWordSet.add(text);
-
-        if(autoCompletionBinding != null)
-            autoCompletionBinding.dispose();
-
-                autoCompletionBinding = TextFields.bindAutoCompletion(searchBar.getEditor(), possibleWordSet);
-    }
+//    private void learnWord(String text) {
+//        possibleWordSet.add(text);
+//
+//        if(autoCompletionBinding != null)
+//            autoCompletionBinding.dispose();
+//
+//                autoCompletionBinding = TextFields.bindAutoCompletion(searchBar.getEditor(), possibleWordSet);
+//    }
 
 
     private Parent nodeInfo;
